[X68030] Fixed NVRAM regression.
git-svn-id: svn://messdev.no-ip.org/mess@9745 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/includes/x68k.h
@@ -42,7 +42,7 @@ public:
 		: driver_device(machine, config),
 		  m_nvram(*this, "nvram") { }
 
-	required_shared_ptr<UINT16>	m_nvram;
+	required_shared_ptr<UINT32>	m_nvram;
 	struct
 	{
 		int sram_writeprotect;
